Added staticProps and semi-authorization
You can get the props at build time including the props name and its default value but will not update on changes. Also props that are objects only go one level deep right now but might add recursion.

Added 1/2 of the semi-authorization not sure if it is actually secure but nuxt do a similar approach not sure if it needs it as well but there when needed.

also added documentation for some helper functions might put them into a separate file later